Move conversion functions to `conv` module
